#commit 7 - DbInitializer refactoring
